Implement Health abstract class, Enemy & Player Health inherits
