+++ isaac2/Assets/Scripts/Health.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6731782b5b36ad2438e525c3f5fa9844
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
