+++ isaac2/Assets/Scripts/PlayerHealth.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerHealth : Health
+{
+    public override void Death()
+    {
+        // Player dies
+    }
+}
